+++ FILES/HEADERS/tuple.h
@@ -1,19 +1,18 @@
-
 #include <stdint.h>
 
 
 typedef struct tuple {
-	uint64_t key ;
-	uint64_t payload ;
+	unsigned long long key ;
+	unsigned long long payload ;
 } tuple;
 
 
 
 
-uint64_t tuple_getkey ( tuple * ) ;
+unsigned long long tuple_getkey ( tuple * ) ;
 
-uint64_t tuple_getpayload ( tuple * ) ;
+unsigned long long tuple_getpayload ( tuple * ) ;
 
-void tuple_setkey ( tuple * , uint64_t ) ;
+void tuple_setkey ( tuple * , unsigned long long ) ;
 
-void tuple_setpayload ( tuple * , uint64_t ) ;
+void tuple_setpayload ( tuple * , unsigned long long ) ;
